further deserialization issue fixed somewhat
Added test cases in StringConverterTests to further test serialization (StringConverter.ToJson)

Added test cases in StringConverterTests to further test StringConverter.ToType and corrected the existing one (did not have removal of quotes which should happen when deserializing)

Added test cases to JsonSerializerOptionsTests for multiple escaped characters and combinations such as escaped character being at the end of a string for both simple string objects and complex objects containing string properties.

Corrected test CanDeserializeInvocationReceiveMessage_05 which had an expected value including quotes for a string that did not have quotes originally in it as "" becomes " and would have been removed during deserialization.

In StringConverter I have made it so that it returns the string without quotes (as expected) when there are no escaped characters present as before if there were no escaped characters there would be an extra set of quotes left around the returned value.  I also inverted the logic on the loop in order to ensure that the loop can exit without appending the letter portion of the escaped character if it is the last character in the string (such as a newline on the end would replace the escaped value as expected but would then attach the 'n' character to the end.

I have replaced the deserialization of escaped characters to correctly replace them when deserializing an object with a string component.  I also added the missing escaped characters that were added to the StringConverter in previous commit.